+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 260303
+
+- Improved Quick start (local workstation) => keycloak config
+
 ## 260227
 
 - Reverted the list and count endpoints for STR to retrieve their own data (`GET /str/activities`, `GET /str/activities/count`) => discuss